refactor(permisson): 리소스 fetch 로직을 store로 분리하여 구조 개선
