+++ apps/core/src/backoffice/domains/users/mapper/permissionMapper.ts
@@ -4,5 +4,5 @@ import { PermissionData } from '../types/permisson';
 export const PermissionMapper = (permission: PermissionResponse): PermissionData => ({
     id: permission.id,
     name: permission.name,
-    permissions: permission.permissions.map(permission => ` ${permission.resourceType}(${permission.resourceIds.sort().toString()}) `),
+    permissions: permission.permissions.map(p => ` ${p.resourceType}(${[...p.resourceIds].sort().toString()}) `),
 });
